dev to master

+++ boost/inception/signals.py
@@ -8,6 +8,14 @@ from .models import Scraper, Images
 
 @receiver(post_save, sender=Scraper)
 def run_spider(sender, instance, created, **kwargs):
+    """
+    runs spider and if success updates its job id
+    :param sender:
+    :param instance:
+    :param created:
+    :param kwargs:
+    :return:
+    """
     if created:
         scrapyd = ScrapydAPI('http://scrapyd:6800')
         job_id = scrapyd.schedule(BOT_NAME, PARSER_NAME)
@@ -18,4 +26,12 @@ def run_spider(sender, instance, created, **kwargs):
 
 @receiver(post_delete, sender=Images)
 def delete_image(sender, instance, **kwargs):
-    os.remove(instance.image.path)
+    """
+    removes image after deleting its instance
+    :param sender:
+    :param instance:
+    :param kwargs:
+    :return:
+    """
+    if os.path.exists(instance.image.path):
+        os.remove(instance.image.path)
